scraper/requirements.txt
selenium
python-dotenv
pytesseract
beautifulsoup4
requests
Pillow
